fix a type annotation

+++ libraries/Microsoft.Bot.Builder.AI.QnA/Models/QnAMakerTraceInfo.cs
@@ -50,7 +50,7 @@ namespace Microsoft.Bot.Builder.AI.QnA
         /// The minimum score threshold, used to filter returned results.
         /// </value>
         [JsonProperty("scoreThreshold")]
-        public double ScoreThreshold { get; set; }
+        public float ScoreThreshold { get; set; }
 
         /// <summary>
         /// Gets or sets number of ranked results that are asked to be returned.
